feat(parse): divide parser logic into modules
