+++ crates/claw_parse/src/parser/statement.rs
@@ -1,0 +1,14 @@
+use super::create_call;
+use crate::ast::{Expression, Statement};
+use chumsky::prelude::*;
+
+pub fn create_statement() -> impl Parser<char, Statement, Error = Simple<char>> {
+    create_call()
+        .then_ignore(just(';').padded())
+        .map(|call| match call {
+            Expression::Call { module, name, args } => Statement::Call { module, name, args },
+            _ => {
+                unreachable!("Tried to convert non-call to a statement, please file a bug report!")
+            }
+        })
+}
